Fix not registering component on startup
Some weirdness is going on. The extension works when the whole
directory is copied to thunderbird user extensions directory and
also works when link file to extension is used. But it didn't work
when it was installed as xpi. It returned error:
While creating services from category 'profile-after-change',
  could not create service for entry
  '@zonio.net/freebusy/provider;1',
  contract ID '2378ee03-7b47-4ae5-9f11-4c41d2ac0b50'
Removing dashes from component name (see diff) helped. But when I
put the dashes back it worked too. And git didn't see any change
in file.
Anyway let's keep dashes out, just in case.